LayoutTests/streams/readablestream-pull-redundant-expected.txt

ReadableStream avoid redundant pull call.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS pullCount is 1
PASS successfullyParsed is true

TEST COMPLETE
